update to dsf.dev namespace
