+++ pom.xml
@@ -14,7 +14,7 @@
 		<compileSource>17</compileSource>
 		<compileTarget>17</compileTarget>
 
-		<dsf.version>0.9.0</dsf.version>
+		<dsf.version>1.0.0-SNAPSHOT</dsf.version>
 		<dsf.location>../dsf</dsf.location>
 	</properties>
 
@@ -38,9 +38,10 @@
 
 	<dependencies>
 		<dependency>
-			<groupId>org.highmed.dsf</groupId>
+			<groupId>dev.dsf</groupId>
 			<artifactId>dsf-bpe-process-base</artifactId>
 			<version>${dsf.version}</version>
+			<scope>provided</scope>
 		</dependency>
 
 		<dependency>
@@ -52,14 +53,14 @@
 
 		<!-- testing -->
 		<dependency>
-			<groupId>org.highmed.dsf</groupId>
+			<groupId>dev.dsf</groupId>
 			<artifactId>dsf-bpe-process-base</artifactId>
 			<version>${dsf.version}</version>
 			<scope>test</scope>
 			<type>test-jar</type>
 		</dependency>
 		<dependency>
-			<groupId>org.highmed.dsf</groupId>
+			<groupId>dev.dsf</groupId>
 			<artifactId>dsf-fhir-validation</artifactId>
 			<version>${dsf.version}</version>
 			<scope>test</scope>
@@ -155,25 +156,6 @@
 					<staticGroups>java,*</staticGroups>
 				</configuration>
 			</plugin>
-			<!--
-			<plugin>
-				<groupId>dev.dsf</groupId>
-				<artifactId>dsf-tools-documentation-generator</artifactId>
-				<version>${dsf.version}</version>
-				<executions>
-					<execution>
-						<goals>
-							<goal>generate</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<workingPackages>
-						<workingPackage>dev.dsf.bpe</workingPackage>
-					</workingPackages>
-				</configuration>
-			</plugin>
-			-->
 		</plugins>
 	</build>
