+++ src/main/java/dev/dsf/bpe/ConstantsUpdateAllowList.java
@@ -1,0 +1,31 @@
+package dev.dsf.bpe;
+
+public interface ConstantsUpdateAllowList
+{
+	String PROCESS_NAME_DOWNLOAD_ALLOW_LIST = "downloadAllowList";
+	String PROCESS_NAME_UPDATE_ALLOW_LIST = "updateAllowList";
+
+	String PROCESS_NAME_FULL_DOWNLOAD_ALLOW_LIST = "dsfdev_" + PROCESS_NAME_DOWNLOAD_ALLOW_LIST;
+	String PROCESS_NAME_FULL_UPDATE_ALLOW_LIST = "dsfdev_" + PROCESS_NAME_UPDATE_ALLOW_LIST;
+
+	String CODESYSTEM_DSF_UPDATE_ALLOW_LIST = "http://dsf.dev/fhir/CodeSystem/update-allow-list";
+	String CODESYSTEM_DSF_UPDATE_ALLOW_LIST_VALUE_ALLOW_LIST = "allow_list";
+
+	String PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST = "http://dsf.dev/fhir/StructureDefinition/task-download-allow-list";
+	String PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST_AND_LATEST_VERSION = PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST + "|"
+			+ UpdateAllowListProcessPluginDefinition.VERSION;
+	String PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI = ConstantsBase.PROCESS_DSF_URI_BASE
+			+ PROCESS_NAME_DOWNLOAD_ALLOW_LIST + "/";
+	String PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI_AND_LATEST_VERSION = PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI
+			+ UpdateAllowListProcessPluginDefinition.VERSION;
+	String PROFILE_DSF_TASK_DOWNLOAD_ALLOW_LIST_MESSAGE_NAME = "downloadAllowListMessage";
+
+	String PROFILE_DSF_TASK_UPDATE_ALLOW_LIST = "http://dsf.dev/fhir/StructureDefinition/task-update-allow-list";
+	String PROFILE_DSF_TASK_UPDATE_ALLOW_LIST_AND_LATEST_VERSION = PROFILE_DSF_TASK_UPDATE_ALLOW_LIST + "|"
+			+ UpdateAllowListProcessPluginDefinition.VERSION;
+	String PROFILE_DSF_TASK_UPDATE_ALLOW_LIST_PROCESS_URI = ConstantsBase.PROCESS_DSF_URI_BASE
+			+ PROCESS_NAME_UPDATE_ALLOW_LIST + "/";
+	String PROFILE_DSF_TASK_UPDATE_ALLOW_LIST_PROCESS_URI_AND_LATEST_VERSION = PROFILE_DSF_TASK_UPDATE_ALLOW_LIST_PROCESS_URI
+			+ UpdateAllowListProcessPluginDefinition.VERSION;
+	String PROFILE_DSF_TASK_UPDATE_ALLOW_LIST_MESSAGE_NAME = "updateAllowListMessage";
+}
