--- src/main/java/org/highmed/dsf/bpe/ConstantsUpdateAllowList.java
@@ -1,34 +1,0 @@
-package org.highmed.dsf.bpe;
-
-import static org.highmed.dsf.bpe.ConstantsBase.PROCESS_HIGHMED_URI_BASE;
-import static org.highmed.dsf.bpe.UpdateAllowListProcessPluginDefinition.VERSION;
-
-public interface ConstantsUpdateAllowList
-{
-	String PROCESS_NAME_DOWNLOAD_ALLOW_LIST = "downloadAllowList";
-	String PROCESS_NAME_UPDATE_ALLOW_LIST = "updateAllowList";
-
-	String PROCESS_NAME_FULL_DOWNLOAD_ALLOW_LIST = "highmedorg_" + PROCESS_NAME_DOWNLOAD_ALLOW_LIST;
-	String PROCESS_NAME_FULL_UPDATE_ALLOW_LIST = "highmedorg_" + PROCESS_NAME_UPDATE_ALLOW_LIST;
-
-	String CODESYSTEM_HIGHMED_UPDATE_ALLOW_LIST = "http://highmed.org/fhir/CodeSystem/update-allow-list";
-	String CODESYSTEM_HIGHMED_UPDATE_ALLOW_LIST_VALUE_ALLOW_LIST = "allow_list";
-
-	String PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST = "http://highmed.org/fhir/StructureDefinition/task-download-allow-list";
-	String PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST_AND_LATEST_VERSION = PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST + "|"
-			+ VERSION;
-	String PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI = PROCESS_HIGHMED_URI_BASE
-			+ PROCESS_NAME_DOWNLOAD_ALLOW_LIST + "/";
-	String PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI_AND_LATEST_VERSION = PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST_PROCESS_URI
-			+ VERSION;
-	String PROFILE_HIGHMED_TASK_DOWNLOAD_ALLOW_LIST_MESSAGE_NAME = "downloadAllowListMessage";
-
-	String PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST = "http://highmed.org/fhir/StructureDefinition/task-update-allow-list";
-	String PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST_AND_LATEST_VERSION = PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST + "|"
-			+ VERSION;
-	String PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST_PROCESS_URI = PROCESS_HIGHMED_URI_BASE
-			+ PROCESS_NAME_UPDATE_ALLOW_LIST + "/";
-	String PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST_PROCESS_URI_AND_LATEST_VERSION = PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST_PROCESS_URI
-			+ VERSION;
-	String PROFILE_HIGHMED_TASK_UPDATE_ALLOW_LIST_MESSAGE_NAME = "updateAllowListMessage";
-}
