+++ src/main/resources/META-INF/services/dev.dsf.bpe.ProcessPluginDefinition
@@ -1,0 +1,1 @@
+dev.dsf.bpe.UpdateAllowListProcessPluginDefinition
