--- src/main/resources/META-INF/services/org.highmed.dsf.bpe.ProcessPluginDefinition
@@ -1,1 +1,0 @@
-org.highmed.dsf.bpe.UpdateAllowListProcessPluginDefinition
