+++ src/main/resources/bpe/downloadAllowList.bpmn
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" id="Definitions_1yb5vw3" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="4.8.1">
-  <bpmn:process id="highmedorg_downloadAllowList" isExecutable="true" camunda:versionTag="#{version}">
+  <bpmn:process id="dsfdev_downloadAllowList" isExecutable="true" camunda:versionTag="#{version}">
     <bpmn:sequenceFlow id="SequenceFlow_0bbhq2r" sourceRef="StartEvent_1" targetRef="downloadAllowListTask" />
     <bpmn:endEvent id="EndEvent_0xd0x8k">
       <bpmn:incoming>SequenceFlow_0oyvmcd</bpmn:incoming>
     </bpmn:endEvent>
     <bpmn:sequenceFlow id="SequenceFlow_0oyvmcd" sourceRef="downloadAllowListTask" targetRef="EndEvent_0xd0x8k" />
-    <bpmn:serviceTask id="downloadAllowListTask" name="downloadAllowList" camunda:class="org.highmed.dsf.bpe.service.DownloadAllowList">
+    <bpmn:serviceTask id="downloadAllowListTask" name="downloadAllowList" camunda:class="dev.dsf.bpe.service.DownloadAllowList">
       <bpmn:incoming>SequenceFlow_0bbhq2r</bpmn:incoming>
       <bpmn:outgoing>SequenceFlow_0oyvmcd</bpmn:outgoing>
     </bpmn:serviceTask>
@@ -17,7 +17,7 @@
   </bpmn:process>
   <bpmn:message id="Message_1nn2wdw" name="downloadAllowListMessage" />
   <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="highmedorg_downloadAllowList">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="dsfdev_downloadAllowList">
       <bpmndi:BPMNEdge id="SequenceFlow_0oyvmcd_di" bpmnElement="SequenceFlow_0oyvmcd">
         <di:waypoint x="365" y="121" />
         <di:waypoint x="415" y="121" />
