--- src/main/resources/fhir/ActivityDefinition/highmed-downloadAllowList.xml
@@ -1,49 +1,0 @@
-<ActivityDefinition xmlns="http://hl7.org/fhir">
-	<meta>
-		<tag>
-			<system value="http://highmed.org/fhir/CodeSystem/read-access-tag" />
-			<code value="ALL" />
-		</tag>
-	</meta>
-	<extension url="http://highmed.org/fhir/StructureDefinition/extension-process-authorization">
-		<extension url="message-name">
-			<valueString value="downloadAllowListMessage" />
-		</extension>
-		<extension url="task-profile">
-			<valueCanonical value="http://highmed.org/fhir/StructureDefinition/task-download-allow-list|#{version}" />
-		</extension>
-		<extension url="requester">
-			<valueCoding>
-				<system value="http://highmed.org/fhir/CodeSystem/process-authorization" />
-				<code value="LOCAL_ALL" />
-			</valueCoding>
-		</extension>
-		<extension url="recipient">
-			<valueCoding>
-				<system value="http://highmed.org/fhir/CodeSystem/process-authorization" />
-				<code value="LOCAL_ALL" />
-			</valueCoding>
-		</extension>
-	</extension>
-	<url value="http://highmed.org/bpe/Process/downloadAllowList" />
-	<!-- version managed by bpe -->
-	<version value="#{version}" />
-	<name value="DownloadAllowList" />
-	<title value="Download Allow List" />
-	<subtitle value="Download Allow List Bundle Process" />
-	<!-- status managed by bpe -->
-	<status value="unknown" />
-	<experimental value="false" />
-	<!-- date managed by bpe -->
-	<date value="#{date}" />
-	<publisher value="HiGHmed" />
-	<contact>
-		<name value="HiGHmed" />
-		<telecom>
-			<system value="email" />
-			<value value="pmo@highmed.org" />
-		</telecom>
-	</contact>
-	<description value="Process to download the allow list FHIR bundle" />
-	<kind value="Task" />
-</ActivityDefinition>
