--- src/main/resources/fhir/ActivityDefinition/highmed-updateAllowList.xml
@@ -1,61 +1,0 @@
-<ActivityDefinition xmlns="http://hl7.org/fhir">
-	<meta>
-		<tag>
-			<system value="http://highmed.org/fhir/CodeSystem/read-access-tag" />
-			<code value="ALL" />
-		</tag>
-	</meta>
-	<extension url="http://highmed.org/fhir/StructureDefinition/extension-process-authorization">
-		<extension url="message-name">
-			<valueString value="updateAllowListMessage" />
-		</extension>
-		<extension url="task-profile">
-			<valueCanonical value="http://highmed.org/fhir/StructureDefinition/task-update-allow-list|#{version}" />
-		</extension>
-		<extension url="requester">
-			<valueCoding>
-				<extension url="http://highmed.org/fhir/StructureDefinition/extension-process-authorization-organization">
-					<valueIdentifier>
-						<system value="http://highmed.org/sid/organization-identifier" />
-						<value value="${org.highmed.dsf.bpe.allow.list.organization:hs-heilbronn.de}" />
-					</valueIdentifier>
-				</extension>
-				<system value="http://highmed.org/fhir/CodeSystem/process-authorization" />
-				<code value="LOCAL_ORGANIZATION" />
-			</valueCoding>
-		</extension>
-		<extension url="recipient">
-			<valueCoding>
-				<extension url="http://highmed.org/fhir/StructureDefinition/extension-process-authorization-organization">
-					<valueIdentifier>
-						<system value="http://highmed.org/sid/organization-identifier" />
-						<value value="${org.highmed.dsf.bpe.allow.list.organization:hs-heilbronn.de}" />
-					</valueIdentifier>
-				</extension>
-				<system value="http://highmed.org/fhir/CodeSystem/process-authorization" />
-				<code value="LOCAL_ORGANIZATION" />
-			</valueCoding>
-		</extension>
-	</extension>
-	<url value="http://highmed.org/bpe/Process/updateAllowList" />
-	<!-- version managed by bpe -->
-	<version value="#{version}" />
-	<name value="UpdateAllowList" />
-	<title value="Update Allow List" />
-	<subtitle value="Update Allow List Bundle Process" />
-	<!-- status managed by bpe -->
-	<status value="unknown" />
-	<experimental value="false" />
-	<!-- date managed by bpe -->
-	<date value="#{date}" />
-	<publisher value="HiGHmed" />
-	<contact>
-		<name value="HiGHmed" />
-		<telecom>
-			<system value="email" />
-			<value value="pmo@highmed.org" />
-		</telecom>
-	</contact>
-	<description value="Process to update the allow list FHIR bundle" />
-	<kind value="Task" />
-</ActivityDefinition>
