--- src/main/resources/fhir/CodeSystem/highmed-update-allow-list.xml
@@ -1,29 +1,0 @@
-<CodeSystem xmlns="http://hl7.org/fhir">
-	<meta>
-		<tag>
-			<system value="http://highmed.org/fhir/CodeSystem/read-access-tag" />
-			<code value="ALL" />
-		</tag>
-	</meta>
-	<url value="http://highmed.org/fhir/CodeSystem/update-allow-list" />
-	<!-- version managed by bpe -->
-	<version value="#{version}" />
-	<name value="HiGHmed_Update_Allow_List" />
-	<title value="HiGHmed Update Allow List" />
-	<!-- status managed by bpe -->
-	<status value="unknown" />
-	<experimental value="false" />
-	<!-- date managed by bpe -->
-	<date value="#{date}" />
-	<publisher value="HiGHmed" />
-	<description value="CodeSystem with standard values for the processes update and download allow list" />
-	<caseSensitive value="true" />
-	<hierarchyMeaning value="grouped-by" />
-	<versionNeeded value="false" />
-	<content value="complete" />
-	<concept>
-		<code value="allow_list" />
-		<display value="Allow List" />
-		<definition value="Allow list identifying organizations" />
-	</concept>
-</CodeSystem>
