--- src/main/resources/fhir/StructureDefinition/highmed-task-update-allow-list.xml
@@ -1,107 +1,0 @@
-<StructureDefinition xmlns="http://hl7.org/fhir">
-  <meta>
-    <tag>
-      <system value="http://highmed.org/fhir/CodeSystem/read-access-tag" />
-      <code value="ALL" />
-    </tag>
-  </meta>
-  <url value="http://highmed.org/fhir/StructureDefinition/task-update-allow-list" />
-  <!-- version managed by bpe -->
-  <version value="#{version}" />
-  <name value="TaskUpdateAllowList" />
-  <!-- status managed by bpe -->
-  <status value="unknown" />
-  <experimental value="false" />
-  <!-- date managed by bpe -->
-  <date value="#{date}" />
-  <fhirVersion value="4.0.1" />
-  <kind value="resource" />
-  <abstract value="false" />
-  <type value="Task" />
-  <baseDefinition value="http://highmed.org/fhir/StructureDefinition/task-base" />
-  <derivation value="constraint" />
-  <differential>
-    <element id="Task.instantiatesUri">
-      <path value="Task.instantiatesUri" />
-      <fixedUri value="http://highmed.org/bpe/Process/updateAllowList/#{version}" />
-    </element>
-    <element id="Task.input">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <max value="1" />
-    </element>
-    <element id="Task.input:message-name">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="message-name" />
-    </element>
-    <element id="Task.input:message-name.value[x]">
-      <path value="Task.input.value[x]" />
-      <fixedString value="updateAllowListMessage" />
-    </element>
-    <element id="Task.input:business-key">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="business-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.input:correlation-key">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="correlation-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.output:allowList">
-      <path value="Task.output" />
-      <sliceName value="allowList" />
-      <max value="1" />
-    </element>
-    <element id="Task.output:allowList.type">
-      <path value="Task.output.type" />
-      <binding>
-        <extension url="http://hl7.org/fhir/StructureDefinition/elementdefinition-bindingName">
-          <valueString value="TaskOutputParameterType" />
-        </extension>
-        <strength value="required" />
-        <valueSet value="http://highmed.org/fhir/ValueSet/update-allow-list|#{version}" />
-      </binding>
-    </element>
-    <element id="Task.output:allowList.type.coding">
-      <path value="Task.output.type.coding" />
-      <min value="1" />
-      <max value="1" />
-    </element>
-    <element id="Task.output:allowList.type.coding.system">
-      <path value="Task.output.type.coding.system" />
-      <min value="1" />
-      <fixedUri value="http://highmed.org/fhir/CodeSystem/update-allow-list" />
-    </element>
-    <element id="Task.output:allowList.type.coding.code">
-      <path value="Task.output.type.coding.code" />
-      <min value="1" />
-      <fixedCode value="allow_list" />
-    </element>
-    <element id="Task.output:allowList.value[x]">
-      <path value="Task.output.value[x]" />
-      <type>
-        <code value="Reference" />
-      </type>
-    </element>
-    <element id="Task.output:allowList.value[x].reference">
-      <path value="Task.output.value[x].reference" />
-      <min value="1" />
-    </element>
-    <element id="Task.output:allowList.value[x].identifier">
-      <path value="Task.output.value[x].identifier" />
-      <max value="0" />
-    </element>
-  </differential>
-</StructureDefinition>
