+++ src/main/resources/fhir/ValueSet/dsf-update-allow-list.xml
@@ -1,0 +1,26 @@
+<ValueSet xmlns="http://hl7.org/fhir">
+	<meta>
+		<tag>
+			<system value="http://dsf.dev/fhir/CodeSystem/read-access-tag" />
+			<code value="ALL" />
+		</tag>
+	</meta>
+	<url value="http://dsf.dev/fhir/ValueSet/update-allow-list" />
+	<!-- version managed by bpe -->
+	<version value="#{version}" />
+	<name value="DSF_Update_Allow_List" />
+	<title value="DSF Update Allow List" />
+	<!-- status managed by bpe -->
+	<status value="unknown" />
+	<experimental value="false" />
+	<!-- date managed by bpe -->
+	<date value="#{date}" />
+	<publisher value="DSF" />
+	<description value="ValueSet with standard values for the process update allow list" />
+	<immutable value="true" />
+	<compose>
+		<include>
+			<system value="http://dsf.dev/fhir/CodeSystem/update-allow-list" />
+		</include>
+	</compose>
+</ValueSet>
