--- src/main/resources/fhir/ValueSet/highmed-update-allow-list.xml
@@ -1,26 +1,0 @@
-<ValueSet xmlns="http://hl7.org/fhir">
-	<meta>
-		<tag>
-			<system value="http://highmed.org/fhir/CodeSystem/read-access-tag" />
-			<code value="ALL" />
-		</tag>
-	</meta>
-	<url value="http://highmed.org/fhir/ValueSet/update-allow-list" />
-	<!-- version managed by bpe -->
-	<version value="#{version}" />
-	<name value="HiGHmed_Update_Allow_List" />
-	<title value="HiGHmed Update Allow List" />
-	<!-- status managed by bpe -->
-	<status value="unknown" />
-	<experimental value="false" />
-	<!-- date managed by bpe -->
-	<date value="#{date}" />
-	<publisher value="HiGHmed" />
-	<description value="ValueSet with standard values for the process update allow list" />
-	<immutable value="true" />
-	<compose>
-		<include>
-			<system value="http://highmed.org/fhir/CodeSystem/update-allow-list" />
-		</include>
-	</compose>
-</ValueSet>
