+++ src/test/resources/log4j2.xml
@@ -6,43 +6,10 @@
 		<Console name="CONSOLE" target="SYSTEM_OUT">
 			<PatternLayout pattern="%p\t%t - %C{1}.%M(%L) | %m%n"/>
 		</Console>
-		<!-- <RollingFile name="FILE" fileName="log/adt_db.log"
-			filePattern="log/adt_db_%d{MM-dd-yyyy}_%i.log.gz"
-			ignoreExceptions="false">
-			<PatternLayout>
-				<Pattern>%d [%t] %-5p %c - %m%n</Pattern>
-			</PatternLayout>
-			<Policies>
-				<OnStartupTriggeringPolicy />
-				<TimeBasedTriggeringPolicy />
-			</Policies>
-		</RollingFile> -->
-
-		<!-- <SMTP name="MAIL" subject="Production RWH LAB Gateway Error"
-			to="hauke.hund@med.uni-heidelberg.de, thomas.hilbel@med.uni-heidelberg.de"
-			from="med3-kvm15@med.uni-heidelberg.de"
-			replyTo="hauke.hund@med.uni-heidelberg.de" smtpHost="localhost"
-			smtpPort="25" bufferSize="50">
-		</SMTP>
-		<SMTP name="MAIL_CERTIFICATE" subject="Production RWH LAB Gateway Certificate Warning"
-			to="hauke.hund@med.uni-heidelberg.de, thomas.hilbel@med.uni-heidelberg.de"
-			from="med3-kvm15@med.uni-heidelberg.de"
-			replyTo="hauke.hund@med.uni-heidelberg.de" smtpHost="localhost"
-			smtpPort="25" bufferSize="1">
-			<ThresholdFilter level="WARN" onMatch="ACCEPT" onMismatch="DENY"/>
-		</SMTP>
-		<SMTP name="MAIL_HL7" subject="Production RWH ADT Gateway HL7 Warning"
-			to="hauke.hund@med.uni-heidelberg.de, thomas.hilbel@med.uni-heidelberg.de"
-			from="med3-kvm15@med.uni-heidelberg.de"
-			replyTo="hauke.hund@med.uni-heidelberg.de" smtpHost="localhost"
-			smtpPort="25" bufferSize="1">
-			<ThresholdFilter level="WARN" onMatch="ACCEPT" onMismatch="DENY"/>
-		</SMTP> -->
 	</Appenders>
 
 	<Loggers>
-		<Logger name="de.rwh" level="TRACE"/>
-		<Logger name="org.highmed" level="TRACE"/>
+		<Logger name="dev.dsf" level="TRACE"/>
 		<Logger name="org.apache" level="WARN"/>
 		<Logger name="org.springframework" level="WARN"/>
 		<Logger name="jndi" level="WARN"/>
@@ -51,17 +18,8 @@
 		<Logger name="liquibase" level="WARN"/>
 		<Logger name="ca.uhn.hl7v2" level="WARN"/>
 
-		<!-- <Logger name="certificate-warning-logger" level="INFO">
-			<AppenderRef ref="MAIL_CERTIFICATE" />
-		</Logger>
-		<Logger name="async-message-handler-warning-logger" level="INFO">
-			<AppenderRef ref="MAIL_HL7" />
-		</Logger> -->
-
 		<Root level="WARN">
 			<AppenderRef ref="CONSOLE"/>
-			<!-- <AppenderRef ref="FILE" /> -->
-			<!-- <AppenderRef ref="MAIL" /> -->
 		</Root>
 	</Loggers>
 </Configuration>
